Adicionando a funcao que calcula o melhor preço e ajustes de layout

+++ app/src/main/java/com/example/calculadoraDeCerveja/Beer.kt
@@ -1,6 +1,6 @@
 package com.example.calculadoraDeCerveja
 
-class Beer(id: Int? = null, marcaBeer: String,  valorBeer: Double,  tamanhoBeer: Double) {
+class Beer(id: Int? = null, marcaBeer: String, valorBeer: Double, tamanhoBeer: Int) {
     var marca = marcaBeer
     var valor = valorBeer
     var tamanho = tamanhoBeer
